requirements.txt
apscheduler
google-api-python-client
groq
matplotlib
numpy
pandas
python-dotenv
requests
streamlit
